motor_sensor_control/requirements.txt
PyQt5==5.15.10
pyserial==3.5
pyqtgraph==0.13.3
numpy==1.24.3
